+++ src/xliffmerge/xlf12-merge.spec.ts
@@ -105,7 +105,7 @@ describe('XliffMerge XLIFF 1.2 format tests', () => {
                     srcDir: WORKDIR,
                     genDir: WORKDIR,
                     i18nFile: MASTERFILE,
-		    useSourceAsTarget: false
+		            useSourceAsTarget: false
                 }
             };
             let xliffMergeCmd = XliffMerge.createFromOptions(commandOut, {languages: ['de']}, profileContent);
@@ -163,7 +163,7 @@ describe('XliffMerge XLIFF 1.2 format tests', () => {
                     srcDir: WORKDIR,
                     genDir: WORKDIR,
                     i18nFile: MASTERFILE,
-		    useSourceAsTarget: false
+		            useSourceAsTarget: false
                 }
             };
             let xliffMergeCmd = XliffMerge.createFromOptions(commandOut, {languages: ['de', 'en']}, profileContent);
@@ -181,6 +181,48 @@ describe('XliffMerge XLIFF 1.2 format tests', () => {
             expect(langFileEnglish.targetLanguage()).toBe('en');
             langFileEnglish.forEachTransUnit((tu: ITransUnit) => {
                 expect(tu.targetContent()).toBe('');
+                expect(tu.targetState()).toBe('new');
+            });
+            done();
+        });
+
+        it('should generate translated file for all languages with set praefix and suffix (#70)', (done) => {
+            FileUtil.copy(MASTER1SRC, MASTER);
+            let ws: WriterToString = new WriterToString();
+            let commandOut = new CommandOutput(ws);
+            let profileContent: IConfigFile = {
+                xliffmergeOptions: {
+                    defaultLanguage: 'de',
+                    srcDir: WORKDIR,
+                    genDir: WORKDIR,
+                    i18nFile: MASTERFILE,
+                    targetPraefix: '%%',
+                    targetSuffix: '!!',
+                }
+            };
+            let xliffMergeCmd = XliffMerge.createFromOptions(commandOut, {languages: ['de', 'en']}, profileContent);
+            xliffMergeCmd.run();
+            expect(ws.writtenData()).not.toContain('ERROR');
+            let langFileGerman: ITranslationMessagesFile = readXliff(xliffMergeCmd.generatedI18nFile('de'));
+            expect(langFileGerman.sourceLanguage()).toBe('de');
+            expect(langFileGerman.targetLanguage()).toBe('de');
+            langFileGerman.forEachTransUnit((tu: ITransUnit) => {
+                if (!tu.targetContent().startsWith('{VAR')) {
+                    expect(tu.targetContent()).toBe('%%' + tu.sourceContent() + '!!');
+                } else {
+                    expect(tu.targetContent()).toBe(tu.sourceContent());
+                }
+                expect(tu.targetState()).toBe('final');
+            });
+            let langFileEnglish: ITranslationMessagesFile = readXliff(xliffMergeCmd.generatedI18nFile('en'));
+            expect(langFileEnglish.sourceLanguage()).toBe('de');
+            expect(langFileEnglish.targetLanguage()).toBe('en');
+            langFileEnglish.forEachTransUnit((tu: ITransUnit) => {
+                if (!tu.targetContent().startsWith('{VAR')) {
+                    expect(tu.targetContent()).toBe('%%' + tu.sourceContent() + '!!');
+                } else {
+                    expect(tu.targetContent()).toBe(tu.sourceContent());
+                }
                 expect(tu.targetState()).toBe('new');
             });
             done();