+++ src/xliffmerge/xmb-merge.spec.ts
@@ -119,6 +119,43 @@ describe('XliffMerge XMB format tests', () => {
             done();
         });
 
+        it('should generate translated file for all languages with set praefix and suffix (#70)', (done) => {
+            FileUtil.copy(MASTER1SRC, MASTER);
+            let ws: WriterToString = new WriterToString();
+            let commandOut = new CommandOutput(ws);
+            let profileContent: IConfigFile = {
+                xliffmergeOptions: {
+                    defaultLanguage: 'de',
+                    srcDir: WORKDIR,
+                    genDir: WORKDIR,
+                    i18nFormat: 'xmb',
+                    i18nFile: MASTERFILE,
+                    targetPraefix: '%%',
+                    targetSuffix: '!!',
+                }
+            };
+            let xliffMergeCmd = XliffMerge.createFromOptions(commandOut, {languages: ['de', 'en']}, profileContent);
+            xliffMergeCmd.run();
+            expect(ws.writtenData()).not.toContain('ERROR');
+            let langFileGerman: ITranslationMessagesFile = readXtbWithMaster(xliffMergeCmd.generatedI18nFile('de'), MASTER);
+            langFileGerman.forEachTransUnit((tu: ITransUnit) => {
+                if (!tu.targetContent().startsWith('{VAR')) {
+                    expect(tu.targetContent()).toBe('%%' + tu.sourceContent() + '!!');
+                } else {
+                    expect(tu.targetContent()).toBe(tu.sourceContent());
+                }
+            });
+            let langFileEnglish: ITranslationMessagesFile = readXtbWithMaster(xliffMergeCmd.generatedI18nFile('en'), MASTER);
+            langFileEnglish.forEachTransUnit((tu: ITransUnit) => {
+                if (!tu.targetContent().startsWith('{VAR')) {
+                    expect(tu.targetContent()).toBe('%%' + tu.sourceContent() + '!!');
+                } else {
+                    expect(tu.targetContent()).toBe(tu.sourceContent());
+                }
+            });
+            done();
+        });
+
         it('should merge translated file for all languages using format xmb', (done) => {
             FileUtil.copy(MASTER1SRC, MASTER);
             let ws: WriterToString = new WriterToString();